app/core/serializers.py
from rest_framework import serializers

from core.models import DigitalProduct, PhysicalProduct


class PhysicalProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = PhysicalProduct
        fields = (
            'name',
            'description',
            'price',
            'stock',
        )

    def validate_price(self, value):
        if value < 0:
            raise serializers.ValidationError("Price cannot be negative")
        return value


class DigitalProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = DigitalProduct
        fields = (
            'name',
            'description',
            'price',
            'stock',
        )

    def validate_price(self, value):
        if value < 0:
            raise serializers.ValidationError("Price cannot be negative")
        return value
